disable E401 and move benchcmarks
