+++ benchmarks/strip_test.py
