+++ tox.ini
@@ -1,4 +1,4 @@
 [flake8]
-ignore = F401,E501,E402
+ignore = F401,E501,E402,E401
 #exclude = tests/*
 max-complexity = -1
